feat: support comments and configurations
This is the first commit of the refactoring task. Highlighting for
comments still works well, but that for configurations (`config.txt`) is
a brand-new one.

We comment all other old language configurations temporarily - that's
really waste of disk space indeed, but we keep them to find out how we
set up those highlighters before quickly. In other words, we don't need
to refer to previous work via GitHub or `git-diff` command.